requirements.txt
# MLBB Discord Bot Dependencies

discord.py>=2.3.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
